refactor: initial commit
